+++ src/providers/formlyConfig.js
@@ -1,4 +1,5 @@
-let angular = require('angular-fix');
+const angular = require('angular-fix');
+const utils = require('../other/utils');
 
 module.exports = ngModule => {
   ngModule.provider('formlyConfig', formlyConfig);
@@ -12,6 +13,10 @@ module.exports = ngModule => {
     var defaultWrapperName = 'default';
     var _this = this;
     var getError = formlyUsabilityProvider.getFormlyError;
+    const allowedTypeProperties = [
+      'name', 'template', 'templateUrl', 'controller', 'link',
+      'defaultOptions', 'extends', 'wrapper', 'data'
+    ];
 
     angular.extend(this, {
       setType,
@@ -37,9 +42,102 @@ module.exports = ngModule => {
         angular.forEach(options, setType);
       } else if (angular.isObject(options)) {
         checkType(options);
+        if (options.extends) {
+          extendTypeOptions(options);
+        }
         typeMap[options.name] = options;
       } else {
         throw getError(`You must provide an object or array for setType. You provided: ${JSON.stringify(arguments)}`);
+      }
+    }
+
+    function checkType(options) {
+      if (!options.name) {
+        throw getError(`You must provide a name for setType. You provided: ${JSON.stringify(arguments)}`);
+      } else if (!options.defaultOptions && !options.template && !options.templateUrl && !options.extends) {
+        throw getError(
+          `You must provide defaultOptions, extends OR a template OR templateUrl for setType. ` +
+          `You provided none of these: ${JSON.stringify(arguments)}`
+        );
+      } else if (options.template && options.templateUrl) {
+        throw getError(
+          `You must provide at most a template OR templateUrl for setType. ` +
+          `You provided both: ${JSON.stringify(arguments)}`
+        );
+      }
+      if (!options.overwriteOk) {
+        checkOverwrite(options.name, typeMap, options, 'types');
+      } else {
+        delete options.overwriteOk;
+      }
+      formlyUsabilityProvider.checkAllowedProperties(allowedTypeProperties, options);
+    }
+
+    function extendTypeOptions(options) {
+      const extendsType = getType(options.extends, true, options);
+      extendTypeControllerFunction(options, extendsType);
+      extendTypeLinkFunction(options, extendsType);
+      extendTypeDefaultOptions(options, extendsType);
+      utils.reverseDeepMerge(options, extendsType);
+    }
+
+    function extendTypeControllerFunction(options, extendsType) {
+      const extendsCtrl = extendsType.controller;
+      if (!angular.isDefined(extendsCtrl)) {
+        return;
+      }
+      const optionsCtrl = options.controller;
+      if (angular.isDefined(optionsCtrl)) {
+        options.controller = function($scope, $controller) {
+          $controller(extendsCtrl, {$scope});
+          $controller(optionsCtrl, {$scope});
+        };
+        options.controller.$inject = ['$scope', '$controller'];
+      } else {
+        options.controller = extendsCtrl;
+      }
+    }
+
+    function extendTypeLinkFunction(options, extendsType) {
+      const extendsFn = extendsType.link;
+      if (!angular.isDefined(extendsFn)) {
+        return;
+      }
+      const optionsFn = options.link;
+      if (angular.isDefined(optionsFn)) {
+        options.link = function() {
+          extendsFn(...arguments);
+          optionsFn(...arguments);
+        };
+      } else {
+        options.link = extendsFn;
+      }
+    }
+
+    function extendTypeDefaultOptions(options, extendsType) {
+      const extendsDO = extendsType.defaultOptions;
+      if (!angular.isDefined(extendsDO)) {
+        return;
+      }
+      const optionsDO = options.defaultOptions;
+      const optionsDOIsFn = angular.isFunction(optionsDO);
+      const extendsDOIsFn = angular.isFunction(extendsDO);
+      if (extendsDOIsFn) {
+        options.defaultOptions = function defaultOptions(options) {
+          var extendsDefaultOptions = extendsDO(options);
+          if (optionsDOIsFn) {
+            return optionsDO(extendsDefaultOptions);
+          } else {
+            utils.reverseDeepMerge(extendsDefaultOptions, optionsDO);
+            return extendsDefaultOptions;
+          }
+        };
+      } else if (optionsDOIsFn) {
+        options.defaultOptions = function defaultOptions(options) {
+          let newDefaultOptions = {};
+          utils.reverseDeepMerge(newDefaultOptions, options, extendsDO);
+          return optionsDO(newDefaultOptions);
+        };
       }
     }
 
@@ -54,27 +152,6 @@ module.exports = ngModule => {
         );
       } else {
         return type;
-      }
-    }
-
-    function checkType(options) {
-      if (!options.name) {
-        throw getError(`You must provide a name for setType. You provided: ${JSON.stringify(arguments)}`);
-      } else if (!options.defaultOptions && !options.template && !options.templateUrl) {
-        throw getError(
-          `You must provide defaultOptions OR a template OR templateUrl for setType. ` +
-          `You provided none of these: ${JSON.stringify(arguments)}`
-        );
-      } else if (options.template && options.templateUrl) {
-        throw getError(
-          `You must provide at most a template OR templateUrl for setType. ` +
-          `You provided both: ${JSON.stringify(arguments)}`
-        );
-      }
-      if (!options.overwriteOk) {
-        checkOverwrite(options.name, typeMap, options, 'types');
-      } else {
-        delete options.overwriteOk;
       }
     }