+++ src/providers/formlyUsability.js
@@ -1,13 +1,15 @@
 var angular = require('angular-fix');
 
 module.exports = ngModule => {
-  ngModule.provider('formlyUsability', function () {
-    var errorsAndWarningsUrlPrefix = 'https://github.com/formly-js/angular-formly/wiki/Errors-and-Warnings#';
+  ngModule.provider('formlyUsability', function(formlyVersion) {
+    var errorsAndWarningsUrlPrefix =
+      `https://github.com/formly-js/angular-formly/blob/${formlyVersion}/other/ERRORS_AND_WARNINGS.md#`;
     angular.extend(this, {
       getFormlyError: getFormlyError,
       getFieldError: getFieldError,
       checkWrapper: checkWrapper,
       checkWrapperTemplate: checkWrapperTemplate,
+      checkAllowedProperties,
       $get: () => this
     });
 
@@ -61,6 +63,23 @@ module.exports = ngModule => {
         );
       }
     }
+
+    function checkAllowedProperties(allowedProperties, obj, context) {
+      var extraProps = Object.keys(obj).filter(prop => allowedProperties.indexOf(prop) === -1);
+      if (extraProps.length) {
+        let extraPropsJSON = JSON.stringify(extraProps.join(', '));
+        let allowedPropsJSON = JSON.stringify(allowedProperties.join(', '));
+        throw getFieldError(
+          'you-have-specified-properties-for-context-that-are-not-allowed',
+          [
+            `You have specified properties for ${context} that are not allowed: ${extraPropsJSON}`,
+            `Allowed properties are: ${allowedPropsJSON}`
+          ].join('\n'),
+          obj
+        );
+      }
+    }
+
 
   });
 };